Improve the implementation (hopefully)
This commit changes to process sets right away (using a queue for mashing purposes). It also changes the poll logic to not process the calculated while we are processing the sets. It also updates the buffer to just return the prediction without the user caring if it's populated. Hopefully this works!